% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-and-write.R
\name{prep_and_write}
\alias{prep_and_write}
\title{Prep and save}
\usage{
prep_and_write(x, save_as, ...)
}
\arguments{
\item{x}{Input dataset returned by rtweet}

\item{save_as}{Name used to save the file. It will be appended with the
numeric range of contained datetimes.}

\item{...}{Named dots along with their coded value are used to add labels
to the dataset prior to writing.}
}
\value{

}
\description{
Prep and save
}
\examples{
## generate example data frame
tweetdata <- data.frame(status_id = c("3459023", "0891370"), text = c("twitter tweet text", "another tweet text"))

## label and save dataset
prep_and_write(tweetdata, "/tmp/tweetdata", test = TRUE)
}
